+++ backend/app/repositories/memory.py
@@ -405,7 +405,7 @@ class MemoryRepository(AbstractRepository):
         return bids
 
     def update_bid_distributed_quantities(
-        self, bid_id: UUID, quantity: int, price_per_unit: Decimal
+        self, bid_id: UUID, quantity: float, price_per_unit: Decimal
     ) -> None:
         """Update the distributed quantity and price for a bid."""
         bid = self._bids.get(bid_id)
